feat(dashboard): split admin dashboard into modular components
